api/requirements.txt
google-cloud-storage
pytz
requests